% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{A map of accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{This is the state wich i want to plot}

\item{years}{This is the year wich i want to recopilate the information and plot}
}
\value{
This function returns a map that indicates the places where happened an
accident
}
\description{
This is a function that give us a map where indicates an accidents locations.
If i select an invalid state num it generate an error and if select a sate num which
does not have any accident a message "no accidents to plot" will appear.
}
\examples{
fars_map_state(1,2014)

}
