% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reading a file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the file i want to read}
}
\value{
This function return the estructure of the file loaded as data frame
}
\description{
This is a simple function that, validate the existence of a file, if the file does
not exist you will see a warning. The file will be read with read_csv function. Then
with the file alreay loaded ii create a data frame.
}
\examples{
fars_read("accident_2014.csv.bz2")

}
