% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{List month-year creator}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{This is the year wich i want the month-year information list}
}
\value{
This function returns a list with month-year information
}
\description{
This is a function that give us a list with the year that our selection(2013,2014 y
2015) we implement the make_filename and fars_read function to get the data.
If we select an unexpected year we'll see a warning message as well we use the
tryCatch function to anticipate the possibility of warning or errors and handle it.
}
\examples{
fars_read_years(2014)

}
