% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{A summary year-month infomation}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{This is the year wich i want the  each month accidents information}
}
\value{
This function returns a table with twelve observations (number of accidents)
for each month
}
\description{
This is a function that give us a summary number of accidents of each month of a
specific year.
}
\examples{
fars_summarize_years(2014)

}
