% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{File name creator}
\usage{
make_filename(year)
}
\arguments{
\item{year}{This is the year of the file i want to generate}
}
\value{
This function returns the file name
}
\description{
This is a simple function that give us the file name with the folowing structure:
"accident_year.csv.bz2".
}
\examples{
make_filename(2013)

}
